feat: replace latest repositories with pinned repositories module
