+++ _includes/component/projects.html
@@ -14,14 +14,14 @@
 
       </div>
 
-      <div id="latest-repos" class="col-xs-12 col-md-9">
+      <div id="pinned-repositories" class="col-xs-12 col-md-9">
         <h3 class="text-white">
-          Latest Repos
+          Pinned Repositories
         </h3>
 
-        <div id="latest-repos-items" class="row">
-          {% include template/latest-repos/placeholder.html %}
-          {% include template/latest-repos/link.html %}
+        <div id="pinned-repositories-items" class="row">
+          {% include template/pinned-repositories/placeholder.html %}
+          {% include template/pinned-repositories/link.html %}
         </div>
 
       </div>
