--- _includes/template/latest-repos/link.html
@@ -1,12 +1,0 @@
-<template id="latest-repos-template">
-  <div class="latest-repos-column col-xs-12 col-md-6 text-md-left mb-3">
-    <a class="latest-repos-link" href="#">
-      <div class="card bg-light hvr-trim h-100 w-100">
-        <div class="card-body">
-          <h5 class="card-title latest-repos-title"></h5>
-          <p class="card-text latest-repos-description"></p>
-        </div>
-      </div>
-    </a>
-  </div>
-</template>
