+++ _includes/template/pinned-repositories/link.html
@@ -1,0 +1,12 @@
+<template id="pinned-repositories-template">
+  <div class="pinned-repositories-column col-xs-12 col-md-6 text-md-left mb-3">
+    <a class="pinned-repositories-link" href="#">
+      <div class="card bg-light hvr-trim h-100 w-100">
+        <div class="card-body">
+          <h5 class="card-title pinned-repositories-title"></h5>
+          <p class="card-text pinned-repositories-description"></p>
+        </div>
+      </div>
+    </a>
+  </div>
+</template>
