+++ _includes/template/pinned-repositories/placeholder.html
@@ -1,6 +1,6 @@
-<template id="latest-repo-placeholder">
-  <div class="latest-repos-link col-xs-12 col-md-6 mb-3">
-      <div class="latest-repo-item panel hvr-trim w-100 h-100">
+<template id="pinned-repositories-placeholder">
+  <div class="pinned-repositories-link col-xs-12 col-md-6 mb-3">
+      <div class="pinned-repositories-item panel hvr-trim w-100 h-100">
         <div class="bg-placeholder rounded" style="overflow: hidden;">
           <div class="placeholder-mask header-top"></div>
           <div class="placeholder-mask header-right"></div>
