+++ _layouts/home.html
@@ -29,8 +29,8 @@ slug: home
       url: '{{site.components.latest_commit.url}}',
       username: '{{site.components.latest_commit.username}}'
     },
-    latestRepos: '{{site.components.latest_repos}}',
     metas: '{{site.components.metas}}',
+    pinnedRepositories: '{{site.components.pinned_repositories}}',
     profiles: '{{site.components.profiles}}',
     quotes: '{{site.components.quotes}}',
     recentlyRead: '{{site.components.recently_read}}'
@@ -39,8 +39,8 @@ slug: home
   window.__WWW_LAYOUT_MODULES__ = [
     'instagram',
     'latest-commit',
-    'latest-repository',
     'navigation',
+    'pinned-repositories',
     'quotes',
     'recently-read',
     'twitter'
