+++ assets/styles/components/_placeholder.scss
@@ -25,7 +25,7 @@
     position: absolute;
 }
 
-.latest-repo-item {
+.pinned-repositories-item {
     .header-top,
     .header-bottom,
     .subheader-bottom {
