+++ assets/styles/components/_projects.scss
@@ -5,7 +5,7 @@
 #projects {
     background-color: $dodger-blue;
 
-    .latest-repos-description {
+    .pinned-repositories-description {
         min-height: 50px;
     }
 }
